+++ freight/__init__.py
@@ -1,4 +1,4 @@
 default_app_config = 'freight.apps.FreightConfig'
 
-__version__ = '1.0.0'
+__version__ = '1.0.1dev1'
 __title__ = 'Alliance Freight'